+++ src/features/auth/components/Login/LoginForm.tsx
@@ -1,13 +1,8 @@
 import React, { useState } from 'react';
 
-import { Container, FormLabel, Input, Text } from '@chakra-ui/react';
-import { useNavigate } from 'react-router-dom';
+import { Container, FormLabel, Input } from '@chakra-ui/react';
 
-import { PATHS } from '@/routes/paths';
 import { FullWidthButton } from '@/styles/theme/components/Buttons';
-import { BlackMarketLogo } from '@/styles/theme/components/Logo';
-
-import { AuthSection } from '../common/AuthSection';
 
 import { LoginFormControl } from './LoginFormControl.styles';
 
@@ -27,8 +22,6 @@ interface CredentialsFormState {
 }
 
 export const LoginForm = () => {
-  const navigate = useNavigate();
-
   const [credentials, setCredentials] = useState<CredentialsFormState>({
     userId: '',
     password: '',
@@ -39,33 +32,20 @@ export const LoginForm = () => {
 
   return (
     <>
-      <AuthSection height="29rem">
-        <BlackMarketLogo width={174} height={31} />
-        <LoginFormControl>
-          {fields.map(({ key, label, ...inputProps }) => (
-            <>
-              <FormLabel htmlFor={key}>{label}</FormLabel>
-              <Input id={key} onChange={handleInputChange} {...inputProps} />
-            </>
-          ))}
-          <FullWidthButton colorScheme="secondary">Log in</FullWidthButton>
-        </LoginFormControl>
-        <Container w="100%" mt="2.75rem">
-          <FullWidthButton colorScheme="lightblue" variant="link">
-            I forgot my password
-          </FullWidthButton>
-        </Container>
-      </AuthSection>
-      <AuthSection height="8rem" mt="1rem" pt="1.5rem">
-        <Text>{`Don't have an account?`}</Text>
-        <FullWidthButton
-          variant="outline"
-          mt="1rem"
-          onClick={() => navigate(`/${PATHS.AUTH}/${PATHS.REGISTER}`)}
-        >
-          Sign up
+      <LoginFormControl>
+        {fields.map(({ key, label, ...inputProps }) => (
+          <>
+            <FormLabel htmlFor={key}>{label}</FormLabel>
+            <Input id={key} onChange={handleInputChange} {...inputProps} />
+          </>
+        ))}
+        <FullWidthButton colorScheme="secondary">Log in</FullWidthButton>
+      </LoginFormControl>
+      <Container w="100%" mt="2.75rem">
+        <FullWidthButton colorScheme="lightblue" variant="link">
+          I forgot my password
         </FullWidthButton>
-      </AuthSection>
+      </Container>
     </>
   );
 };
