+++ src/features/auth/components/Login/index.ts
@@ -1,1 +1,1 @@
-export * from './LoginForm';
+export * from './LoginFormContainer';
